# src/main/java/org/example/app/entity/book/Book.java
package org.example.app.entity.book;

import com.fasterxml.jackson.annotation.JsonIgnore;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import lombok.*;
import org.example.app.entity.author.Author;
import org.example.app.entity.book.review.BookReview;
import org.example.app.entity.genre.Genre;

import javax.persistence.*;
import java.time.LocalDate;
import java.util.List;
import org.example.app.entity.tag.Tag;
import org.example.app.entity.user.UserEntity;


@Data
@AllArgsConstructor
@NoArgsConstructor

@Entity
@Table(name = "book")
@ApiModel(description = "entity representing a book")
public class Book {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(columnDefinition = "INT NOT NULL AUTO_INCREMENT")
    @ApiModelProperty("id generated by db automatically")
    private Integer id;

    @Column(columnDefinition = "TEXT")
    private String description;

    @Column(columnDefinition = "SMALLINT NOT NULL DEFAULT 0")
//  скидка в процентах или 0, если её нет
    private Integer discount;

    @Column(columnDefinition = "VARCHAR(255)")
//  изображение обложки
    private String image;

    @Column(name = "is_bestseller",
            columnDefinition = "SMALLINT NOT NULL")
//  книга очень популярна, является бестселлером
    private Boolean isBestseller;

    @Column(columnDefinition = "INT NOT NULL")
//  цена в рублях основная
    private Integer price;

    @Column(name = "pub_date", columnDefinition = "DATE NOT NULL")
//  дата публикации
    private LocalDate pubDate;

    @Column(columnDefinition = "VARCHAR(255) NOT NULL")
//  мнемонический идентификатор книги
    private String slug;

    @Column(name= "number_users_bought_book", columnDefinition = "INT")
//  количество пользователей, купивших книгу
    private Integer numberUsersBoughtBook;

    @Column(name= "number_users_book_in_card", columnDefinition = "INT")
//  количество пользователей, у которых книга находится в корзине;
    private Integer numberUsersBookInCard;

    @Column(columnDefinition = "INT")
//  количество пользователей, у которых книга находится отложена;
    private Integer numberUsersBookPostponed;

    @Column(columnDefinition = "VARCHAR(255) NOT NULL")
    @ApiModelProperty("book title")
    private String title;

    @ManyToMany
    @JoinTable(name = "book2author",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "author_id"))
    @JsonIgnore
    private List<Author> authors;

    @ManyToMany
    @JoinTable(name = "book2genre",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "genre_id"))
    @JsonIgnore
    private List<Genre> genres;

    @ManyToMany
    @JoinTable(name = "book2tag",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "tag_id"))
    @JsonIgnore
    private List<Tag> tags;

    @ManyToMany
    @JoinTable(name = "book2user",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "user_id"))
    @JsonIgnore
    private List<UserEntity> users;

    @ManyToMany
    @JoinTable(name = "file_download",
            joinColumns = {@JoinColumn(name = "book_id")},
            inverseJoinColumns = {@JoinColumn(name = "user_id")})
    @JsonIgnore
    private List<UserEntity> usersDownloadedBooks;

    @ManyToMany
    @JoinTable(name = "balance_transaction",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "user_id"))
    @JsonIgnore
    private List<UserEntity> usersBalanceTransactions;

    @ManyToMany
    @JoinTable(name = "book_review",
            joinColumns = @JoinColumn(name = "book_id"),
            inverseJoinColumns = @JoinColumn(name = "user_id"))
    @JsonIgnore
    private List<UserEntity> usersBookReviews;

    @OneToMany(mappedBy = "book")
    @JsonIgnore
    private List<BookReview> bookReviewList;
}